+++ Rebus.Tests/Integration/TestRetryExceptionCustomization.cs
@@ -11,7 +11,6 @@ using Rebus.Routing.Exceptions;
 using Rebus.Tests.Contracts;
 using Rebus.Tests.Contracts.Extensions;
 using Rebus.Tests.Contracts.Utilities;
-using Rebus.Tests.Extensions;
 using Rebus.Transport.InMem;
 #pragma warning disable 1998