+++ Rebus/Retry/ErrorTracking/InMemErrorTracker.cs
@@ -56,7 +56,8 @@ namespace Rebus.Retry.ErrorTracking
                 id => new ErrorTracking(exception),
                 (id, tracking) => tracking.AddError(exception));
 
-            _log.Warn("Unhandled exception {errorNumber} while handling message with ID {messageId}: {exception}", errorTracking.Errors.Count(), messageId, exception);
+            _log.Warn("Unhandled exception {errorNumber} while handling message with ID {messageId}", errorTracking.Errors.Count(), messageId);
+            //_log.Warn(exception, "Unhandled exception {errorNumber} while handling message with ID {messageId}", errorTracking.Errors.Count(), messageId);
         }
 
         /// <summary>