+++ Rebus/Retry/PoisonQueues/PoisonQueueErrorHandler.cs
@@ -24,12 +24,9 @@ namespace Rebus.Retry.PoisonQueues
         /// </summary>
         public PoisonQueueErrorHandler(SimpleRetryStrategySettings simpleRetryStrategySettings, ITransport transport, IRebusLoggerFactory rebusLoggerFactory)
         {
-            if (simpleRetryStrategySettings == null) throw new ArgumentNullException(nameof(simpleRetryStrategySettings));
-            if (transport == null) throw new ArgumentNullException(nameof(transport));
-            if (rebusLoggerFactory == null) throw new ArgumentNullException(nameof(rebusLoggerFactory));
-            _simpleRetryStrategySettings = simpleRetryStrategySettings;
-            _transport = transport;
-            _log = rebusLoggerFactory.GetLogger<PoisonQueueErrorHandler>();
+            _log = rebusLoggerFactory?.GetLogger<PoisonQueueErrorHandler>() ?? throw new ArgumentNullException(nameof(rebusLoggerFactory));
+            _transport = transport ?? throw new ArgumentNullException(nameof(transport));
+            _simpleRetryStrategySettings = simpleRetryStrategySettings ?? throw new ArgumentNullException(nameof(simpleRetryStrategySettings));
         }
 
         /// <summary>
@@ -43,7 +40,7 @@ namespace Rebus.Retry.PoisonQueues
         /// <summary>
         /// Handles the poisonous message by forwarding it to the configured error queue
         /// </summary>
-        public async Task HandlePoisonMessage(TransportMessage transportMessage, ITransactionContext transactionContext, string errorDescription)
+        public async Task HandlePoisonMessage(TransportMessage transportMessage, ITransactionContext transactionContext, Exception exception)
         {
             var headers = transportMessage.Headers;
 
@@ -54,20 +51,20 @@ namespace Rebus.Retry.PoisonQueues
                 messageId = "<unknown>";
             }
 
-            headers[Headers.ErrorDetails] = errorDescription;
+            headers[Headers.ErrorDetails] = exception.ToString();
             headers[Headers.SourceQueue] = _transport.Address;
 
             var errorQueueAddress = _simpleRetryStrategySettings.ErrorQueueAddress;
 
             try
             {
-                _log.Error("Moving message with ID {messageId} to error queue {queueName} - reason: {errorDetails}", messageId, errorQueueAddress, errorDescription);
+                _log.Error(exception, "Moving message with ID {messageId} to error queue {queueName}", messageId, errorQueueAddress);
 
                 await _transport.Send(errorQueueAddress, transportMessage, transactionContext);
             }
-            catch (Exception exception)
+            catch (Exception forwardException)
             {
-                _log.Error(exception, "Could not move message with ID {messageId} to error queue {queueName} - will pause {pauseInterval} to avoid thrashing",
+                _log.Error(forwardException, "Could not move message with ID {messageId} to error queue {queueName} - will pause {pauseInterval} to avoid thrashing",
                     messageId, errorQueueAddress, MoveToErrorQueueFailedPause);
 
                 // if we can't move to error queue, we need to avoid thrashing over and over